+++ src/Intermediate/Exercise1/SolutionES/Multp.java
@@ -1,18 +1,18 @@
-package Intermediate.Exercise1;
+package Intermediate.Exercise1.SolutionES;
 
 /**
  *
  * @author dvchinx
  */
-public class Resta {
+public class Multp {
 
     protected float a, b;
 
-    public Resta() {
+    public Multp() {
     }
 
     public int Operacion(float a, float b) {
-        return (int) Math.round(a - b);
+        return (int) Math.round(a * b);
     }
 
     public float getA() {
